Fix create listing bug. Refactor controller.
