+++ client/react/components/Navbar/index.jsx
@@ -51,7 +51,7 @@ const NavBar = ({
     let viewListings;
 
     if (user && user.isLandlord) {
-        createListing = <Menu.Item onClick={() => navigateTo('/create-listing')}>Create Listing</Menu.Item>
+        createListing = <Menu.Item onClick={() => navigateTo('/create-listing')}><Icon name='add'/>Create Listing</Menu.Item>
         viewListings = <Menu.Item onClick={() => navigateTo('/listings')}><Icon name='home'/>Listings</Menu.Item>
     }
